refactor chat adapter tests
