+++ qabelbox/src/test/java/de/qabel/qabelbox/chat/view/adapters/TextChatMessageViewHolderTest.kt
@@ -1,4 +1,4 @@
-package de.qabel.qabelbox.chat.view.adapter
+package de.qabel.qabelbox.chat.view.adapters
 
 import android.view.View
 import android.widget.LinearLayout
